Rename lib files and imports
