+++ app/lib/check-env.ts
